+++ confidential-assets/tests/units/api/deposit.test.ts
@@ -1,37 +1,23 @@
-import {
-  confidentialAsset,
-  getTestAccount,
-  longTestTimeout,
-  mintFungibleTokens,
-  sendAndWaitTx,
-  TOKEN_ADDRESS,
-} from "../../helpers";
-import { preloadTables } from "../../helpers/wasmPollardKangaroo";
+import { aptos, confidentialAsset, getTestAccount, TOKEN_ADDRESS } from "../../helpers";
 
 describe("Deposit", () => {
   const alice = getTestAccount();
 
-  it(
-    "Pre load wasm table map",
-    async () => {
-      await preloadTables();
-    },
-    longTestTimeout,
-  );
-
   const DEPOSIT_AMOUNT = 5n;
   it("it should deposit Alice's balance of fungible token to her confidential balance", async () => {
-    await mintFungibleTokens(alice);
+    await aptos.fundAccount({
+      accountAddress: alice.accountAddress,
+      amount: 100000000,
+    });
 
     const depositTx = await confidentialAsset.deposit({
-      sender: alice.accountAddress,
+      signer: alice,
       tokenAddress: TOKEN_ADDRESS,
       amount: DEPOSIT_AMOUNT,
     });
-    const resp = await sendAndWaitTx(depositTx, alice);
 
-    console.log("gas used:", resp.gas_used);
+    console.log("gas used:", depositTx.gas_used);
 
-    expect(resp.success).toBeTruthy();
+    expect(depositTx.success).toBeTruthy();
   });
 });