+++ src/api/transaction.ts
@@ -34,6 +34,7 @@ import {
   AnyRawTransaction,
   InputGenerateTransactionOptions,
   InputGenerateTransactionPayloadData,
+  InputTransactionPluginData,
 } from "../transactions";
 import { AccountAddressInput, AuthenticationKey, Ed25519PrivateKey } from "../core";
 import { Account } from "../account";
@@ -522,6 +523,7 @@ export class Transaction {
   async rotateAuthKey(
     args: {
       fromAccount: Account;
+      options?: InputGenerateTransactionOptions;
     } & (
       | { toAccount: Account; dangerouslySkipVerification?: never }
       | { toNewPrivateKey: Ed25519PrivateKey; dangerouslySkipVerification?: never }
@@ -717,7 +719,7 @@ export class Transaction {
     args: FeePayerOrFeePayerAuthenticatorOrNeither & {
       signer: Account;
       transaction: AnyRawTransaction;
-    },
+    } & InputTransactionPluginData,
   ): Promise<PendingTransactionResponse> {
     return signAndSubmitTransaction({
       aptosConfig: this.config,
@@ -744,11 +746,13 @@ export class Transaction {
    * @return PendingTransactionResponse
    * @group Transaction
    */
-  async signAndSubmitAsFeePayer(args: {
-    feePayer: Account;
-    senderAuthenticator: AccountAuthenticator;
-    transaction: AnyRawTransaction;
-  }): Promise<PendingTransactionResponse> {
+  async signAndSubmitAsFeePayer(
+    args: {
+      feePayer: Account;
+      senderAuthenticator: AccountAuthenticator;
+      transaction: AnyRawTransaction;
+    } & InputTransactionPluginData,
+  ): Promise<PendingTransactionResponse> {
     return signAndSubmitAsFeePayer({
       aptosConfig: this.config,
       ...args,