+++ src/types/types.ts
@@ -1,6 +1,8 @@
 // Copyright © Aptos Foundation
 // SPDX-License-Identifier: Apache-2.0
 
+import type { AptosConfig } from "../api";
+import { InputSubmitTransactionData } from "../transactions";
 import { Network } from "../utils/apiEndpoints";
 import { OrderBy, TokenStandard } from "./indexer";
 
@@ -157,6 +159,21 @@ export enum AnyPublicKeyVariant {
   FederatedKeyless = 4,
 }
 
+export function anyPublicKeyVariantToString(variant: AnyPublicKeyVariant): string {
+  switch (variant) {
+    case AnyPublicKeyVariant.Ed25519:
+      return "ed25519";
+    case AnyPublicKeyVariant.Secp256k1:
+      return "secp256k1";
+    case AnyPublicKeyVariant.Keyless:
+      return "keyless";
+    case AnyPublicKeyVariant.FederatedKeyless:
+      return "federated_keyless";
+    default:
+      throw new Error("Unknown public key variant");
+  }
+}
+
 /**
  * Variants of signature types used for cryptographic operations.
  */
@@ -256,6 +273,8 @@ export type AptosSettings = {
   readonly faucetConfig?: FaucetConfig;
 
   readonly transactionGenerationConfig?: TransactionGenerationConfig;
+
+  readonly pluginSettings?: PluginSettings;
 };
 
 /**
@@ -338,6 +357,73 @@ export type TransactionGenerationConfig = {
 export type ClientHeadersType = {
   HEADERS?: Record<string, string | number | boolean>;
 };
+
+/**
+ * Config for plugins. This can be used to override certain client behavior.
+ */
+export type PluginConfig = {
+  /**
+   * If given, this will be used for submitting transactions instead of the default
+   * implementation (which submits transactions directly via a node).
+   */
+  TRANSACTION_SUBMITTER?: TransactionSubmitter;
+
+  /**
+   * If true, we won't use the TRANSACTION_SUBMITTER if set.
+   */
+  IGNORE_TRANSACTION_SUBMITTER?: boolean;
+};
+
+export type PluginSettings = Omit<PluginConfig, "IGNORE_TRANSACTION_SUBMITTER">;
+
+/**
+ * You can implement this interface and set it in {@link PluginSettings} when building a
+ * client to override the default transaction submission behavior. This is useful if
+ * you'd like to submit transactions via a gas station for example.
+ *
+ * @example
+ * ```typescript
+ * class MyGasStationClient implements TransactionSubmitter {
+ *   async submitTransaction(
+ *     args: { aptosConfig: AptosConfig } & InputSubmitTransactionData,
+ *   ): Promise<PendingTransactionResponse> {
+ *     // TODO: Implement the logic to submit the transaction to the gas station
+ *   }
+ * }
+ *
+ * const network = Network.MAINNET;
+ * const myGasStationClient = new MyGasStationClient(network);
+ * const config = new AptosConfig({
+ *   network,
+ *   pluginConfig: {
+ *     transactionSubmitter: myGasStationClient,
+ *   },
+ * });
+ * const aptos = new Aptos(config);
+ * ```
+ */
+export interface TransactionSubmitter {
+  /**
+   * Submit a transaction to the Aptos blockchain or something that will do it on your
+   * behalf, for example a gas station. See the comments of {@link TransactionSubmitter} for more.
+   *
+   * @param args - The arguments for submitting the transaction.
+   * @param args.aptosConfig - The configuration for connecting to the Aptos network.
+   * @param args.transaction - The Aptos transaction data to be submitted.
+   * @param args.senderAuthenticator - The account authenticator of the transaction sender.
+   * @param args.secondarySignerAuthenticators - Optional. Authenticators for additional signers in a multi-signer transaction.
+   * @param args.pluginParams - Optional. Additional parameters for the plugin.
+   * @param args.transactionSubmitter - Optional. An override for the transaction submitter.
+   *
+   * @returns PendingTransactionResponse - The response containing the status of the submitted transaction.
+   * @group Implementation
+   */
+  submitTransaction(
+    args: {
+      aptosConfig: AptosConfig;
+    } & Omit<InputSubmitTransactionData, "transactionSubmitter">,
+  ): Promise<PendingTransactionResponse>;
+}
 
 /**
  * Represents a client for making requests to a service provider.
@@ -1002,7 +1088,8 @@ export type TransactionSignature =
   | TransactionSecp256k1Signature
   | TransactionMultiEd25519Signature
   | TransactionMultiAgentSignature
-  | TransactionFeePayerSignature;
+  | TransactionFeePayerSignature
+  | TransactionSingleSenderSignature;
 
 /**
  * Determine if the provided signature is an Ed25519 signature.
@@ -1012,8 +1099,8 @@ export type TransactionSignature =
  * @param signature - The transaction signature to be checked.
  * @returns A boolean indicating whether the signature is an Ed25519 signature.
  */
-export function isEd25519Signature(signature: TransactionSignature): signature is TransactionFeePayerSignature {
-  return "signature" in signature && signature.signature === "ed25519_signature";
+export function isEd25519Signature(signature: TransactionSignature): signature is TransactionEd25519Signature {
+  return "signature" in signature && signature.type === "ed25519_signature";
 }
 
 /**
@@ -1059,6 +1146,18 @@ export function isMultiEd25519Signature(
 }
 
 /**
+ * Determine if the provided signature is of type "single_sender".
+ *
+ * @param signature - The transaction signature to check.
+ * @returns A boolean indicating whether the signature is a single-sender signature.
+ */
+export function isSingleSenderSignature(
+  signature: TransactionSignature,
+): signature is TransactionSingleSenderSignature {
+  return signature.type === "single_sender";
+}
+
+/**
  * The signature for a transaction using the Ed25519 algorithm.
  */
 export type TransactionEd25519Signature = {
@@ -1074,6 +1173,15 @@ export type TransactionSecp256k1Signature = {
   type: string;
   public_key: string;
   signature: "secp256k1_ecdsa_signature";
+};
+
+/**
+ * The structure for a multi-signature transaction using Ed25519.
+ */
+export type TransactionSingleSenderSignature = {
+  type: "single_sender";
+  public_key: { value: string; type: string };
+  signature: { value: string; type: string };
 };
 
 /**